requirements.txt
fastapi
uvicorn
requests
python-dotenv
anthropic
openai
pyyaml
PyPDF2
